Move unimplemented status command to respective componenet files and bump rdk to merged commit
